+++ react-native/storyly-react-native/StorylyDemo/App.js
@@ -11,7 +11,10 @@
 import React, { Component, useState, useRef, useEffect } from 'react';
 import { View, Image, Text, PixelRatio, ScrollView, TouchableOpacity, Animated } from 'react-native';
 import { Storyly } from 'storyly-react-native';
-
+import { VerticalFeedBar } from 'storyly-react-native';
+import { VerticalFeed } from 'storyly-react-native';
+import { VerticalFeedPresenter } from 'storyly-react-native';
+import { Dimensions } from 'react-native'
 
 const PIN_ICON = require('./assets/pin_icon.png');
 
@@ -23,7 +26,7 @@ export default class App extends Component {
         super(props);
 
         this.state = {
-            isHidden: true,
+            isHidden: false,
         };
     }
 
@@ -31,19 +34,32 @@ export default class App extends Component {
         return (
             <View>
                 <ScrollView>
-                    <ShowHideAnimation style={{ width: "100%", height: 100 }} isHidden={this.state.isHidden} >
-                        <Storyly
-                            style={{ width: '100%', height: 100, backgroundColor: "#00ffff" }}
-                            storylyId={STORYLY_TOKEN}
-                            storyGroupSize="small"
+                    <ShowHideAnimation style={{ width: "100%", height: Dimensions.get('window').height - 50 }} isHidden={this.state.isHidden} >
+                        <VerticalFeedPresenter
+                            ref={ref => { this.verticalFeedPresenter = ref }}
+                            style={{ width: '100%', height: Dimensions.get('window').height, backgroundColor: "#00ffff" }}
+                            storylyId={"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhY2NfaWQiOjIzODAsImFwcF9pZCI6MjA1MTEsImluc19pZCI6MjI5NDJ9.TXCs-M6guskLJA1JXmu7PlmPxUKfyw88lBpOdxmgpDI"}
                             onLoad={loadEvent => {
-                                console.log(`[Storyly] default - onLoad`);
-                                this.setState({isHidden: false})
+                                console.log(`[Vertical Feed Presenter] - onLoad`);
+                                this.verticalFeedPresenter.resumeStory()
+                            }}
+                            onEvent={ verticalFeedEvent => {
+                                console.log(`[Vertical Feed Presenter] - onEvent ${verticalFeedEvent.event} ${verticalFeedEvent.feedGroup.toString()}`);
+                            }}
+                            onCartUpdate={ StoryProductCartUpdateEvent => {
+                                console.log(`[Vertical Feed Presenter] - onCartUpdate ${StoryProductCartUpdateEvent.event} ${StoryProductCartUpdateEvent.change.toString()}`);
+                            }}
+                            onUserInteracted={ VerticalFeedItemInteractiveEvent => {
+                                console.log(`[Vertical Feed Presenter] - onUserInteracted ${VerticalFeedItemInteractiveEvent.storyComponent}`);
+                            }}
+                            onPress={ VerticalFeedPressEvent => {
+                                console.log(`[Vertical Feed Presenter] - onPress ${VerticalFeedPressEvent.feedItem}`);
                             }} />
                     </ShowHideAnimation>
+
                     <TouchableOpacity onPress={() => { this.setState({isHidden: !this.state.isHidden}) }}>
                         <View style={{ padding: 10, backgroundColor: 'blue' }}>
-                            <Text style={{ color: 'white', textAlign: "center" }}>{this.state.isHidden ? 'Show' : 'Hide'}</Text>
+                            <Text style={{ color: 'white', textAlign: "center" }}>{this.state.isHidden ? 'Show Presenter' : 'Hide Presenter'}</Text>
                         </View>
                     </TouchableOpacity>
 
@@ -83,15 +99,49 @@ export default class App extends Component {
                         storyGroupIconBorderColorSeen={["#FFFFFF", "#FFFFFF"]}
                         storyGroupIconBackgroundColor={"#000000"}
                         storyGroupPinIconColor={"#000000"} />
-                    <Storyly
-                        ref={ref => { this.storyly = ref }}
-                        style={{ width: '100%', height: 178, marginTop: 10, backgroundColor: "#ff00ff" }}
-                        storylyId={STORYLY_TOKEN}
-                        storyGroupViewFactory={{
-                            width: convertToNative(100),
-                            height: convertToNative(178),
-                            customView: CustomPortraitView
-                        }} />
+                    <VerticalFeedBar
+                        ref={ref => { this.verticalFeedBar = ref }}
+                        verticalFeedGroupIconHeight={Platform.OS === 'android' ? PixelRatio.getPixelSizeForLayoutSize(200) : 200}
+                        style={{ width: '100%', height: 250, marginTop: 10, backgroundColor: "#ff00ff" }}
+                        onLoad={loadEvent => {
+                            console.log(`[Vertical Feed Bar] - onLoad`);
+                        }}
+                        onEvent={ verticalFeedEvent => {
+                            console.log(`[Vertical Feed Bar] - onEvent ${verticalFeedEvent.event} ${verticalFeedEvent.feedGroup.toString()}`);
+                        }}
+                        onCartUpdate={ StoryProductCartUpdateEvent => {
+                            console.log(`[Vertical Feed Bar] - onCartUpdate ${StoryProductCartUpdateEvent.event} ${StoryProductCartUpdateEvent.change.toString()}`);
+                        }}
+                        onUserInteracted={ VerticalFeedItemInteractiveEvent => {
+                            console.log(`[Vertical Feed Bar] - onUserInteracted ${VerticalFeedItemInteractiveEvent.storyComponent}`);
+                        }}
+                        onPress={ VerticalFeedPressEvent => {
+                            console.log(`[Vertical Feed Bar] - onPress ${VerticalFeedPressEvent.feedItem}`);
+                        }}
+                        storyGroupIconCornerRadius={Platform.OS === 'android' ? PixelRatio.getPixelSizeForLayoutSize(0) : 0}
+                        storylyId={"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhY2NfaWQiOjIzODAsImFwcF9pZCI6MjA1MTEsImluc19pZCI6MjI5NDJ9.TXCs-M6guskLJA1JXmu7PlmPxUKfyw88lBpOdxmgpDI"}/>
+                    <VerticalFeed
+                        ref={ref => { this.verticalFeed = ref }}
+                        verticalFeedGroupListSections={2}
+                        verticalFeedGroupIconHeight={Platform.OS === 'android' ? PixelRatio.getPixelSizeForLayoutSize(200) : 200}
+                        style={{ width: '100%', height: 750, marginTop: 10}}
+                        onLoad={loadEvent => {
+                            console.log(`[Vertical Feed] - onLoad`);
+                        }}
+                        onEvent={ verticalFeedEvent => {
+                            console.log(`[Vertical Feed] - onEvent ${verticalFeedEvent.event} ${verticalFeedEvent.feedGroup.toString()}`);
+                        }}
+                        onCartUpdate={ StoryProductCartUpdateEvent => {
+                            console.log(`[Vertical Feed] - onCartUpdate ${StoryProductCartUpdateEvent.event} ${StoryProductCartUpdateEvent.change.toString()}`);
+                        }}
+                        onUserInteracted={ VerticalFeedItemInteractiveEvent => {
+                            console.log(`[Vertical Feed] - onUserInteracted ${VerticalFeedItemInteractiveEvent.storyComponent}`);
+                        }}
+                        onPress={ VerticalFeedPressEvent => {
+                            console.log(`[Vertical Feed] - onPress ${VerticalFeedPressEvent.feedItem}`);
+                        }}
+                        storyGroupIconCornerRadius={Platform.OS === 'android' ? PixelRatio.getPixelSizeForLayoutSize(0) : 0}
+                        storylyId={"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhY2NfaWQiOjIzODAsImFwcF9pZCI6MjA1MTEsImluc19pZCI6MjI5NDJ9.TXCs-M6guskLJA1JXmu7PlmPxUKfyw88lBpOdxmgpDI"}/>
                 </ScrollView>
             </View>
         );
@@ -147,3 +197,7 @@ const ShowHideAnimation = (props) => {
         </Animated.View>
     );
 };
+
+Object.prototype.toString = function(){
+    return JSON.stringify(this)
+}