+++ web/RELEASENOTES.md
@@ -1,5 +1,14 @@
 # Release Notes
 
+### 3.6.27 - 12.01.2025 [for cdn](https://web-story.storyly.io/sdk/3.6.27/storyly-web.js)
+
+- Improvement: Added localization support for "Total" and "Go to checkout" texts
+
+### 3.6.26 - 10.01.2025 [for cdn](https://web-story.storyly.io/sdk/3.6.26/storyly-web.js)
+
+- Improvement: If sales price and price the same, price is not showing
+- Bug: Fixed bar layout oveflow after using scroll indicator
+
 ### 3.6.24 - 17.10.2024 [for cdn](https://web-story.storyly.io/sdk/3.6.24/storyly-web.js)
 
 - Bug: Fixed cta, swipe up and sponsored text colors on dark mode